requirements.txt
asgiref==3.5.1
Cython==0.29.28
dango==0.0.1
dartsclone==0.9.0
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
pygtrie==2.4.2
pytz==2022.1
sortedcontainers==2.1.0
sqlparse==0.4.2
SudachiDict-core==20211220
SudachiPy==0.5.4
tzdata==2022.1
requests~=2.27.1